Move bit ops to separate modules
